refactor: reorganize the requestAnimationFrame mocks to be under the general tests folder for reusing it
