+++ src/components/SplitButton/SplitButtonMenu/__tests__/splitButtonMenu-snapshot-tests.jest.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import renderer from "react-test-renderer";
 import SplitButtonMenu from "../SplitButtonMenu";
 import MenuItem from "../../../Menu/MenuItem/MenuItem";
-import { mockRequestAnimationFrame, restoreRequestAnimationFrameMock } from "./test-utils";
+import { mockRequestAnimationFrame, restoreRequestAnimationFrameMock } from "../../../../tests/__tests__/test-utils";
 
 const SplitButtonMenuWithItems = (
   <SplitButtonMenu id="menu">
